requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
instructor>=0.4.0
openai>=1.3.0
azure-ai-documentintelligence>=1.0.0b1
azure-core>=1.29.0
azure-identity>=1.15.0
pandas>=2.1.0
openpyxl>=3.1.0
xlrd>=2.0.1
tabulate>=0.9.0
python-dotenv>=1.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
python-magic>=0.4.27
pandas
openpyxl
xlrd
azure-ai-documentintelligence
azure-identity
python-magic
tabulate


--extra-index-url https://test.pypi.org/simple/
azure-ai-projects
azure-ai-agents==1.1.0b4
azure-identity